# Optimize Multi-head Latent Attention (MLA) with Fast Path for Short Sequences