@a-holm 2025-04-04
The `MLA.forward` method has been significantly refactored. The original code had distinct logic for `attn_impl="naive"` and `attn_impl="absorb"`. The new code uses a unified `matmul`-based approach in the standard path.

Could you confirm that this refactoring preserves the exact behavior of the original code for *both* `naive` and `absorb` attention implementations when the fast path isn't used (i.e., `seqlen > 256` or `mask` is present)? Specifically, how is the logic previously handled in the `absorb` path (involving `weight_dequant` and specific `einsum` operations) now covered?